+++ NEWS.md
@@ -1,3 +1,14 @@
+# 1.5.0
+
+- New LabbcatEdit functions
+  + *versionInfo* - gets version information of all components of LaBB-CAT
+  + *saveParticipant* - add or update a participant record
+  + *saveTranscript* - update transcript attributes
+- New LabbcatView function
+  + *readOnlyCategories* - lists attribute categories for read-only users
+
+Requires LaBB-CAT version 20230508.1434
+
 # 1.4.0
 
 - Changed LabbcatView functions